refactor(chat): remove unused editor association logic from attachment widgets
